codex: fix blob-size-policy on PR #16282
